+++ dither_go/__init__.py
@@ -5,4 +5,5 @@ __version__ = "0.2.1"
 
 from .wrapper import *
 from .matrices import *
+from .utils.matrices import *
 from .exceptions import DitherGoError, InvalidColorError